move enums into minersc
