+++ code/go/0chain.net/smartcontract/minersc/enums/globals.go
